+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Reformatted demo website to horizontal layout.
 -   Added more documentation for the `--output` flag.
 -   Added zoom functionality to output canvas.
+-   Disabled download buttons when there is no input/output.
 
 ### 🔧 Internal changes